requirements.txt
Flask
Pyrebase4
gunicorn
